moving the watson api packages to use python 3.9 for now
