+++ Pipfile
@@ -37,5 +37,11 @@ flask = ">=3.0.0"
 sqlalchemy = "<2.1"
 psycopg2-binary = "*"
 
+[connector-api-packages]
+flask = ">=3.0.0"
+pydantic = ">=2.10.5"
+ibm-watson = ">=9.0.0"
+ibm-cloud-sdk-core = "3.22.1"
+
 [requires]
 python_version = "3.9"
