+++ common/auth.py
@@ -4,6 +4,11 @@ import requests
 from common.config import app
 from .header import Header
 
+from flask import request, jsonify
+import base64
+import functools
+import json
+
 import jwt
 
 from rasa_sdk import Tracker
@@ -76,3 +81,27 @@ def _jwt_decode(token: str) -> None:
         token,
         options={"verify_signature": False, "verify_exp": True, "verify_nbf": True},
     )
+
+
+def check_identity(identity_header):
+    try:
+        base64.b64decode(identity_header)
+    except Exception:
+        return False
+    return True
+
+def require_identity_header(func):
+    @functools.wraps(func)
+    def wrapper(*args, **kwargs):
+        identity_header = request.headers.get('x-rh-identity')
+        if not identity_header or not check_identity(identity_header):
+            return jsonify(message="Invalid x-rh-identity"), 401
+        return func(*args, **kwargs)
+    return wrapper
+
+def get_org_id_from_identity(identity):
+    decoded_identity = base64.b64decode(identity).decode("utf8")
+    identity_json = json.loads(decoded_identity)
+    identity = identity_json.get("identity", {})
+    org_id = identity.get("org_id")
+    return org_id
