--- watson/Pipfile
@@ -1,16 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-flask = ">=3.0.0"
-pydantic = "*"
-ibm-watson = "*"
-ibm-cloud-sdk-core = "*"
-python-decouple = "*"
-
-[requires]
-python_version = "3.12"
-
-[dev-packages]
