--- watson/common/config/__init__.py
@@ -1,9 +1,0 @@
-import os
-
-from collections import ChainMap
-from decouple import Config
-
-__repository_chain = [os.environ]
-
-
-config = Config(ChainMap(*__repository_chain))
